Initial commit to Github. 2011.138.0 release.
The prior home for this code was in RubyForge as best I can tell, and
that's gone now. I've extracted this code from the gem at RubyGems,
though I probably have a local setup of it on an older Windows machine
around my house. Still, this should be fine to get it put into github
where I can do some 2.0 compatibility coding on it should I need.