Fix for File.h_size(filename)
This was added loooong ago and never used. I think I took it from a ruby
mailing list post. I decided to use it recently and realized the last
line of it, formatting it, was incorrect, presuming there would be
decimals. But numbers like 700MB were formatted as 7MB - clearly wrong.

Another method was refactored out as well, so `File.to_h_size(size)` can
be used for any number, and not tied to a specific file.

What is in there now is close enough for my usage. This gem appears to
do a more thorough job of it. On Rails, there's also the
`number_to_human_size` method in ActionView.